Refactor FIU mode changes
