+++ kpf/ddoi_configurations/kpf_inst_config.ini
@@ -44,7 +44,7 @@ guider_xpa_target = CRED2
 [times]
 ao_pcu_shim_time = 5
 ao_pcu_move_time = 150
-fiu_mode_shim_time = 10
+fiu_mode_shim_time = 1
 fiu_fold_mirror_move_time = 40
 fiu_hatch_move_time = 2
 fvc_command_timeout = 5
